Add `OutputResult`
